requirements.txt
asgiref==3.7.2
Django==4.2.3
pycryptodome==3.18.0
sqlparse==0.4.4
typing_extensions==4.7.1
